Use fixed template for basis of expected data
Writing a generator that would work on all cases - especially invalid
cases is cumbersome and does run the risk of 'hiding' bugs - it isn't
uncome for a test writer to parallel bugs in application code in test
code. On the other hand, I am still lazy so I kept some generation code
that can be run and used as a basis for other things in the tests.

There is still one test that deliberately fails in this set - the
missing VNIC type. It is basically a required field so its absence
should not be ignored. I think an exception in convert_vif_to_env is a
valid approach.